+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Added
 
+- Added `esma_postinstall.cmake` script for tarring up code post install
+
 ## [3.6.2] - 2021-Oct-06
 
 ### Changed